+++ changelog.md
@@ -47,3 +47,8 @@
     1. implemented routing for product list
     2. created page for product description
     3. implemented routing for cart
+    4. added components orders
+    5. implemented admin panel with CanActivate guard
+    6. added resolve guard to product-form component
+    7. applied lazy loading to admin module
+    8. changed cart service to store cart items in the local storage